Refactor pagination settings into single class

+++ shim-server/src/main/java/org/openmhealth/shimmer/common/domain/parameters/NumberRequestParameter.java
@@ -25,7 +25,7 @@ import java.util.Optional;
  * Todo: Decide the type for this class
  * @author Chris Schaefbauer
  */
-public class NumberRequestParameter extends RequestParameter<Double> {
+public class NumberRequestParameter extends RequestParameter<BigDecimal> {
 
     private BigDecimal maximumValue;
     private Double minimumValue;
